requirement.txt
sentence-transformers==2.6.1
transformers==4.38.2
faiss-cpu==1.13.2
torch==2.2.1
PyPDF2==3.0.1
numpy==1.26.4
